webapp/requirements.txt
absl-py==0.9.0
annoy==1.16.3
astunparse==1.6.3
boto3==1.12.49
botocore==1.15.49
Brotli==1.0.7
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dash==1.11.0
dash-bootstrap-components==0.9.2
dash-core-components==1.9.1
dash-html-components==1.0.3
dash-renderer==1.4.0
dash-table==4.6.2
dataclasses==0.7
docutils==0.15.2
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gast==0.3.3
google-auth==1.14.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
joblib==0.14.1
Keras-Preprocessing==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
nltk==3.5
numpy==1.18.3
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.3
plotly==4.6.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.41
scipy==1.4.1
sentencepiece==0.1.86
six==1.14.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-hub==0.8.0
termcolor==1.1.0
tokenizers==0.5.2
tqdm==4.45.0
transformers==2.8.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
